Added Support for Terasic/Arrow DECA (10M50) board
This board appears to have been discontinued, but arrow were recently
selling it off for cheap so there's probably a few out there that will
appreciate being able to use this toolchain against it.

As there is no UART available on the board (all examples provided with
the board use the UART-JTAG IP), I have assigned it to gpio0_0 and
gpio0_1 (of the BBB header) for now and a usb-uart will need to be
attached to those pins.